Anusha-Risk-Management-World-Tracker/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
celery==5.3.4
redis==5.0.1
yfinance==0.2.32
requests==2.31.0
reportlab==4.0.7
pandas==2.1.3
numpy==1.26.2
email-validator==2.1.0
